api: add maximum storage space for user

+++ MyCloudStore.Service/DataLayer/Models/AppUser.cs
@@ -11,7 +11,7 @@ namespace MyCloudStore.Service.DataLayer.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
 
-        public int MaxKBs { get; set; }
+        public long MaxKBs { get; set; }
 
         public IList<File> Files { get; set; }
     }
